shared users tasks won't change,delete or update.

+++ server/routes/deleteTask.js
@@ -6,6 +6,9 @@ const authenticateToken = require("../middlewares/authenticateToken");
 router.put("/", authenticateToken, async (req, res) => {
   try {
     if (req.body.userData.signedIn === true) {
+      if (req.body.taskId == undefined) {
+        return(res.status(400).json("You cannot delete another user's task."));
+      };
       const query = 'DELETE FROM `db_tasks` WHERE `task_id` = ?';
       const value = [req.body.taskId];
       await database(query, value);
